# config/swagger.config.yaml
openapi: 3.0.1
info:
    title: BlueShop API
    description: 'This is a Backend API of BlueShop'
    termsOfService: https://api.fashionshop.nkt2001.tech/terms/
    contact:
        email: ktuan439@gmail.com
    version: 1.0.0
servers:
    - url: http://localhost:5000/api/v1
    - url: https://api.fashionshop.nkt2001.tech/api/v1
...
paths:
    /banner:
        get:
            summary: Get all banner and slider (non-user, user, admin)
            description: Get all banner and slider (non-user, user, admin)
            tags:
                - Banner
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        default: true
                                    message:
                                        type: string
                                    data:
                                        type: object
                                        properties:
                                            banners:
                                                type: array
                                                items:
                                                    $ref: '#/components/schemas/Banner'
                                            sliders:
                                                type: array
                                                items:
                                                    $ref: '#/components/schemas/Banner'
        post:
            description: Create a new banner or new slider
            summary: Create a new banner or new slider
            tags:
                - Banner
            security:
                - bearerAuth: []
            requestBody:
                description: Payload to create a new banner or new slider
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BannerForm'

            responses:
                '201':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        default: true
                                    message:
                                        type: string
                                    data:
                                        type: object
                                        properties:
                                            newBanner:
                                                $ref: '#/components/schemas/Banner'

                '400':
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /banner/{id}:
        get:
            summary: Get all banner and slider (non-user, user, admin)
            description: Get all banner and slider (non-user, user, admin)
            tags:
                - Banner
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  description: Banner ID
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        default: true
                                    message:
                                        type: string
                                    data:
                                        type: object
                                        properties:
                                            banner:
                                                $ref: '#/components/schemas/Banner'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'

        put:
            summary: Update exsited Banner
            tags:
                - Banner
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  description: Banner ID
                  required: true
                  schema:
                      type: string
            requestBody:
                description: Payload to update a new banner
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BannerForm'
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Banner'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            summary: Delete existed banner
            description: Delete existed banner
            security:
                - bearerAuth: []
            tags:
                - Banner
            parameters:
                - name: id
                  in: path
                  description: Banner ID
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                '404':
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        # /banner/{id}/restore:
        #     patch:
        #         summary: Restore disabled Banner
        #         description: Restore disabled Banner
        #         security:
        #             - bearerAuth: []
        #         tags:
        #             - Banner
        #         parameters:
        #             - name: id
        #               in: path
        #               description: Banner ID
        #               required: true
        #               schema:
        #                   type: string
        #         requestBody:
        #             description: isDisabled value
        #             required: true
        #             content:
        #                 application/json:
        #                     schema:
        #                         type: object
        #                         properties:
        #                             isDisabled:
        #                                 type: boolean
        #                                 description: Is disabled or not
        #                                 nullable: false
        #                                 default: false
        #         responses:
        #             '200':
        #                 description: Success
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Banner'
        #             '404':
        #                 description: Not Found
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        #             '401':
        #                 description: Unauthorized
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        # /banner/{id}/disable:
        #     patch:
        #         summary: Disable a Banner
        #         description: Disable a Banner
        #         security:
        #             - bearerAuth: []
        #         tags:
        #             - Banner
        #         parameters:
        #             - name: id
        #               in: path
        #               description: Banner ID
        #               required: true
        #               schema:
        #                   type: string
        #         requestBody:
        #             description: isDisabled value
        #             required: true
        #             content:
        #                 application/json:
        #                     schema:
        #                         type: object
        #                         properties:
        #                             isDisabled:
        #                                 type: boolean
        #                                 description: Is disabled or not
        #                                 nullable: false
        #                                 default: true
        #         responses:
        #             '200':
        #                 description: Success
        #                 content:
        #                     application/json:
        #                         schema:
        #                             type: object
        #                             properties:
        #                                 message:
        #                                     type: string
        #             '404':
        #                 description: Not Found
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        #             '400':
        #                 description: Bad request
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        #             '401':
        #                 description: Unauthorized
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
...
    /user:
        post:
            summary: Create new user account
            description: This can only be done by the logged in user.
            tags:
                - User
            requestBody:
                description: Payload to create new user
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    nullable: false
                                email:
                                    type: string
                                    nullable: false
                                phone:
                                    type: string
                                    nullable: false
                                password:
                                    type: string
                                    nullable: false
            responses:
                '201':
                    description: Success
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/UserProfile'
                                    - type: object
                                      properties:
                                          refreshToken:
                                              type: string
                                              description: JWT Refresh Token
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        get:
            summary: Get all users by admin
            description: Get all users by admin
            tags:
                - User
            security:
                - bearerAuth: []
            parameters:
                - in: query
                  name: dateOrder
                  required: false
                  schema:
                      type: string
                  description: Sort users by date (newest/latest)
                - in: query
                  name: status
                  required: false
                  schema:
                      type: string
                  description: Filter users by status (disabled/not_disabled/all)
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    allOf:
                                        - $ref: '#/components/schemas/User'
                                        - type: object
                                          properties:
                                              createdAt:
                                                  type: string
                                                  description: The date and time when this record is created
                                                  format: date-time
                                              updatedAt:
                                                  type: string
                                                  description: The date and time when this record is updated
                                                  format: date-time
                                              __v:
                                                  type: number
                                                  description: Document version key
                                              password:
                                                  type: string
                                                  description: User password
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
components:
    schemas:
        Address:
            type: object
            properties:
                province:
                    type: string
                    description: Province
                district:
                    type: string
                    description: District
                ward:
                    type: string
                    description: Ward
                specificAddress:
                    type: string
                    description: Specific Address
        Banner:
            type: object
            properties:
                _id:
                    type: string
                    description: Banner Id
                title:
                    type: string
                    description: Banner title  or slider title
                index:
                    type: number
                    description: Banner index or slider index
                imageUrl:
                    type: string
                    description: URL image
                linkTo:
                    type: string
                    description: Link to URL
                role:
                    type: string
                    description: Role is banner or slider
                disabled:
                    type: boolean
                    description: Banner or slider is Disabled or not
                    default: false
        BannerForm:
            type: object
            required:
                - title
                - index
                - linkTo
                - role
            properties:
                title:
                    type: string
                    description: Banner title or slider title
                index:
                    type: number
                    description: Banner index or slider index
                # bannerImage:
                #     type: file
                #     format: binary
                #     description: Banner image (select file or fill in imageUrl in the form below)
                imageUrl:
                    type: string
                    description: URL image
                linkTo:
                    type: string
                    description: Link to URL
                role:
                    type: string
                    description: Role is 'banner' or 'slider'
                    default: banner
...
        User:
            type: object
            properties:
                _id:
                    type: string
                    description: user id
                name:
                    type: string
                    description: username
                email:
                    type: string
                    description: email
                phone:
                    type: string
                    description: phone
                avatar:
                    type: string
                    description: avatar url
                role:
                    type: string
                    description: user role
                gender:
                    type: string
                    description: user gender
                birthday:
                    type: string
                    description: user birthday
                address:
                    $ref: '#/components/schemas/Address'
                discountCode:
                    type: array
                    description: Discount code list
                isVerified:
                    type: boolean
                    description: user is verify account
                disabled:
                    type: boolean
                    description: is Disabled or not
        UserProfile:
            type: object
            properties:
                _id:
                    type: string
                    description: user id
                name:
                    type: string
                    description: username
                email:
                    type: string
                    description: email
                phone:
                    type: string
                    description: phone
                avatarUrl:
                    type: string
                    description: avatar url
                role:
                    type: string
                    description: user role
                gender:
                    type: string
                    description: user sex
                birthday:
                    type: string
                    description: user birthday
                discountCode:
                    type: array
                    description: Discount code list
                address:
                    $ref: '#/components/schemas/Address'
        Error:
            type: object
            properties:
                success:
                    type: boolean
                    default: false
                message:
                    type: string
                    description: error message
                errors:
                    type: array
                    description: error list
                    default: []
                stack:
                    type: string
                    description: error message description
...
    securitySchemes:
        bearerAuth: # arbitrary name for the security scheme
            type: apiKey
            description: "<i>JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.</i> <b> Example</b>: <i><u>Bearer 12345abcdef</u></i>"
            in: header # can be "header", "query" or "cookie"
            name: Authorization # name of the header, query parameter or cookie
    # 2) Apply the security globally to all operations
...